add modal alert for deleting booking

+++ lib/user/deleteBooking.ts
@@ -1,19 +1,77 @@
-"use server"
+"use server";
 
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
 import prisma from "../db";
 import { revalidatePath } from "next/cache";
+import { actionClient } from "./safe-action";
+import { redirect } from "next/navigation";
+import { z } from "zod";
 
-export async function deleteBooking(bookingId: string) {
-  try {
-    await prisma.booking.delete({
-      where: {
-        id: bookingId,
-      },
-    });
-      revalidatePath("/book-now");
-      return {succeess:"Booking successfully deleted"}
-  } catch (error) {
-    console.error("Failed to delete booking:", error);
-    throw new Error("Failed to delete booking");
-  }
-}
+// export async function deleteBooking(bookingId: string) {
+//   try {
+//     const booking = await prisma.booking.findUnique({
+//       where: {
+//         id: bookingId,
+//       },
+//     });
+
+//     if (!booking) {
+//       return { error: "Booking does not exist" };
+//     }
+
+//     await prisma.booking.delete({
+//       where: {
+//         id: bookingId,
+//       },
+//     });
+
+//     revalidatePath("/track");
+//     return { success: "Booking successfully deleted" };
+//   } catch (error) {
+//     console.error("Failed to delete booking:", error);
+//     return { error: "Failed to delete booking" };
+//   }
+// }
+
+const deleteSchema = z.object({
+  bookingId: z.string(),
+});
+
+export const deleteBooking = actionClient
+  .schema(deleteSchema)
+  .action(async ({ parsedInput: { bookingId } }) => {
+    const { isAuthenticated, getUser } = getKindeServerSession();
+
+    const [isUserAuthenticated] = await Promise.all([
+      isAuthenticated,
+      getUser(),
+    ]);
+
+    if (!isUserAuthenticated) {
+      redirect("/api/auth/login?");
+    }
+
+    try {
+      const booking = await prisma.booking.findUnique({
+        where: {
+          id: bookingId,
+        },
+      });
+
+      if (!booking) {
+        return { error: "Booking does not exist" };
+      }
+
+      await prisma.booking.delete({
+        where: {
+          id: bookingId,
+        },
+      });
+
+      revalidatePath("/track");
+      return { success: "Booking successfully deleted" };
+    } catch (error) {
+      console.error("Failed to delete booking:", error);
+      return { error: "Failed to delete booking" };
+    }
+  });
